:hammer: Refactor serialization/deserialization classes into SerDes namespace
