+++ lib/factorix.rb
@@ -1,15 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "factorix/deserializer"
 require_relative "factorix/errors"
 require_relative "factorix/mod"
 require_relative "factorix/mod_list"
 require_relative "factorix/mod_settings"
 require_relative "factorix/runtime"
-require_relative "factorix/serializer"
+require_relative "factorix/ser_des"
 require_relative "factorix/version"
-require_relative "factorix/version24"
-require_relative "factorix/version64"
 
 # Factorix is a Factorio mod management and development tool.
 module Factorix
