+++ lib/factorix/ser_des.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Factorix
+  # Serialization/Deserialization utilities
+  module SerDes
+  end
+end
+
+require_relative "ser_des/deserializer"
+require_relative "ser_des/serializer"
+require_relative "ser_des/version24"
+require_relative "ser_des/version64"
