+++ sig/factorix/ser_des.rbs
@@ -1,0 +1,5 @@
+module Factorix
+  # Serialization/Deserialization utilities
+  module SerDes
+  end
+end
